+++ RobotCleaner/Models/Shared/ExecutionReport.cs
@@ -9,7 +9,7 @@ namespace RobotCleaner.Models.Shared;
 /// <param name="Commands">The number of commands executed</param>
 /// <param name="Result">The number of unique locations visited</param>
 /// <param name="Duration">The length of time taken to execute the commands</param>
-public record ExecutionReport(DateTime Timestamp, int Commands, int Result, double Duration)
+public record ExecutionReport(DateTime Timestamp, int Commands, long Result, double Duration)
 {
     /// <summary>
     /// The unique ID, which is only populated after the execution report has been written to the database.
